+++ demo/src/main/java/com/rtcc/demo/controller/ProfessorController.java
@@ -4,15 +4,19 @@ import com.rtcc.demo.DTOs.ProfessorRequestDTO;
 import com.rtcc.demo.DTOs.ProfessorResponseDTO;
 import com.rtcc.demo.model.Professor;
 import com.rtcc.demo.repository.ProfessorRepository;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/professor")
-
+@Tag(name = "Professor", description = "API to manage professors")
 public class ProfessorController {
 
     @Autowired
@@ -20,14 +24,22 @@ public class ProfessorController {
 
     @CrossOrigin(origins = "*", allowedHeaders = "*")
     @PostMapping
-    public void saveProfessor(@RequestBody ProfessorRequestDTO data) {
+    @Operation(summary = "Sava a new professor", description = "Add a new professor into database", responses = {
+            @ApiResponse(responseCode = "200", description = "Professor saved successfully")
+    })
+    public ResponseEntity<Void> saveProfessor(@RequestBody ProfessorRequestDTO data) {
         Professor professorData = new Professor(data);
         professorRepository.save(professorData);
-        ResponseEntity.ok().build();
+        return ResponseEntity.ok().build();
     }
 
     @CrossOrigin(origins = "*", allowedHeaders = "*")
     @GetMapping
+    @Operation(summary = "Returns a list of all professors",
+            description = "",
+            responses = {
+                    @ApiResponse(responseCode = "200", description = "List of all professors")
+            })
     public List<ProfessorResponseDTO> getAll() {
         List<ProfessorResponseDTO> professorList = professorRepository
                 .findAll()
@@ -39,15 +51,35 @@ public class ProfessorController {
 
     @CrossOrigin(origins = "*", allowedHeaders = "*")
     @GetMapping("/{id}")
-    public ProfessorResponseDTO getProfessor(@PathVariable String id) {
-        Professor professor = professorRepository.findById(id).orElseThrow();
-        return new ProfessorResponseDTO(professor);
+    @Operation(summary = "Get professor by ID",
+            description = "Get a professor by its ID, if it exists. Otherwise, return 404 Not Found.",
+            responses = {
+                    @ApiResponse(responseCode = "200", description = "Success"),
+                    @ApiResponse(responseCode = "404", description = "Professor not found")
+            })
+    public ResponseEntity<ProfessorResponseDTO> getProfessor(@PathVariable String id) {
+        Optional<Professor> professor = professorRepository.findById(id);
+
+        if (professor.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        ProfessorResponseDTO professorResponse = new ProfessorResponseDTO(professor.get());
+        return ResponseEntity.ok().body(professorResponse);
     }
 
     @CrossOrigin(origins = "*", allowedHeaders = "*")
     @DeleteMapping("/{id}")
-    public void deleteProfessor(@PathVariable String id) {
-        professorRepository.deleteById(id);
-        ResponseEntity.ok().build();
+    @Operation(summary = "Remove professor by id", description = "Remove a professor by its ID, if it exists. Otherwise, return 404 Not Found.",
+            responses = {
+                    @ApiResponse(responseCode = "204", description = "Professor removed successfully"),
+                    @ApiResponse(responseCode = "404", description = "Professor not found")
+            })
+    public ResponseEntity<Void> deleteProfessor(@PathVariable String id) {
+        if (professorRepository.existsById(id)) {
+            professorRepository.deleteById(id);
+            return ResponseEntity.noContent().build(); // Retornar 204 No Content
+        } else {
+            return ResponseEntity.notFound().build(); // Retornar 404 Not Found
+        }
     }
 }
